#29 Made menu more redable and added opion to copy booking instance to template.

+++ report.php
@@ -191,6 +191,12 @@ if ($action == 'downloadpdf') {
     die();
 }
 
+if ($action == 'copytotemplate' && has_capability('mod/booking:manageoptiontemplates', $context) &&
+         confirm_sesskey()) {
+    $bookingdata->copytotemplate();
+    redirect($baseurl, get_string('copytotemplatesucesfull', 'booking'), 5);
+}
+
 if ($action == 'deletebookingoption' && $confirm == 1 &&
          has_capability('mod/booking:updatebooking', $context) && confirm_sesskey()) {
              $bookingdata->delete_booking_option();
